Fix the issue #198(utf-8 multi-byte encoded char is not supported)
detail:
1. fix the wrong unicode converting part and also the insert position(length in std::string) logic(this part done by @djowel)
2. add a TextEdit with chinese char(the attached font is open source, and widely used in GNU/Linux, no license concern) to ease the demo/verify

+++ lib/src/support/text_utils.cpp
@@ -46,7 +46,7 @@ namespace cycfi { namespace elements
    {
       char const* codepoint_to_utf8(unsigned cp, char str[8])
       {
-         int n = 0;
+         size_t n = 0;
          if (cp < 0x80) n = 1;
          else if (cp < 0x800) n = 2;
          else if (cp < 0x10000) n = 3;
